+++ src/pages/profile-orders/profile-orders.tsx
@@ -1,7 +1,7 @@
 import { ProfileOrdersUI } from '@ui-pages';
 import { FC, useEffect } from 'react';
 import { useDispatch, useSelector } from '../../services/store';
-import { getOrders, selectOrders } from '../../services/slices/feedSlice';
+import { getOrders, selectOrders } from '../../services/slices/userSlice';
 
 export const ProfileOrders: FC = () => {
   const { data } = useSelector(selectOrders);
