+++ src/services/store.ts
@@ -8,11 +8,17 @@ import {
 
 import userReducer from './slices/userSlice';
 import feedReducer from './slices/feedSlice';
+import ingredientsReducer from './slices/ingredientsSlice';
+import orderDetailsReducer from './slices/orderDetailsSlice';
+import constructorReducer from './slices/constructorSlice';
 
 const store = configureStore({
   reducer: {
     user: userReducer,
-    feed: feedReducer
+    feed: feedReducer,
+    ingredients: ingredientsReducer,
+    orderDetails: orderDetailsReducer,
+    constructorItems: constructorReducer
   },
   devTools: process.env.NODE_ENV !== 'production'
 });
